feat: Post一覧コンポーネントを汎用的に利用できるように修正する
